Remove TreeNodeV2 and collaps TreeNode interface into only implementation
