+++ community/index/src/main/java/org/neo4j/index/bptree/TreeNode.java
@@ -24,92 +24,383 @@ import java.util.function.Consumer;
 
 import org.neo4j.io.pagecache.PageCursor;
 
-public interface TreeNode<KEY,VALUE>
+/**
+ * Methods to manipulate single node such as set and get header fields,
+ * insert and fetch keys, values and children.
+ *
+ * DESIGN
+ *
+ * Using Separate design the internal nodes should look like
+ *
+ * # = empty space
+ *
+ * [                    HEADER               ]|[      KEYS     ]|[     CHILDREN      ]
+ * [TYPE][KEYCOUNT][RIGHTSIBLING][LEFTSIBLING]|[[KEY][KEY]...##]|[[CHILD][CHILD]...##]
+ *  0     1         5             13            21
+ *
+ * Calc offset for key i (starting from 0)
+ * HEADER_LENGTH + i * SIZE_KEY
+ *
+ * Calc offset for child i
+ * HEADER_LENGTH + SIZE_KEY * MAX_KEY_COUNT_INTERNAL + i * SIZE_CHILD
+ *
+ *
+ * Using Separate design the leaf nodes should look like
+ *
+ *
+ * [                   HEADER                ]|[      KEYS     ]|[       VALUES      ]
+ * [TYPE][KEYCOUNT][RIGHTSIBLING][LEFTSIBLING]|[[KEY][KEY]...##]|[[VALUE][VALUE]...##]
+ *  0     1         5             13            21
+ *
+ * Calc offset for key i (starting from 0)
+ * HEADER_LENGTH + i * SIZE_KEY
+ *
+ * Calc offset for value i
+ * HEADER_LENGTH + SIZE_KEY * MAX_KEY_COUNT_LEAF + i * SIZE_VALUE
+ *
+ * @param <KEY> type of key
+ * @param <VALUE> type of value
+ */
+class TreeNode<KEY,VALUE>
 {
-    void initializeLeaf( PageCursor cursor );
+    private static final int BYTE_POS_TYPE = 0;
+    private static final int BYTE_POS_KEYCOUNT = 1;
+    private static final int BYTE_POS_RIGHTSIBLING = 5;
+    private static final int BYTE_POS_LEFTSIBLING = 13;
+    private static final int HEADER_LENGTH = 21;
 
-    void initializeInternal( PageCursor cursor );
+    private static final int SIZE_CHILD = Long.BYTES;
 
-    boolean isLeaf( PageCursor cursor );
+    private static final byte LEAF_FLAG = 1;
+    private static final byte INTERNAL_FLAG = 0;
+    private static final long NO_NODE_FLAG = -1;
 
-    boolean isInternal( PageCursor cursor );
+    private final int internalMaxKeyCount;
+    private final int leafMaxKeyCount;
+    private final Layout<KEY,VALUE> layout;
 
-    int keyCount( PageCursor cursor );
+    private final int keySize;
+    private final int valueSize;
 
-    long rightSibling( PageCursor cursor );
+    TreeNode( int pageSize, Layout<KEY,VALUE> layout )
+    {
+        this.layout = layout;
+        this.keySize = layout.keySize();
+        this.valueSize = layout.valueSize();
+        this.internalMaxKeyCount = Math.floorDiv( pageSize - (HEADER_LENGTH + SIZE_CHILD), keySize + SIZE_CHILD);
+        this.leafMaxKeyCount = Math.floorDiv( pageSize - HEADER_LENGTH, keySize + valueSize );
+    }
 
-    long leftSibling( PageCursor cursor );
+    // ROUTINES
 
-    void setTypeLeaf( PageCursor cursor );
+    void initializeLeaf( PageCursor cursor )
+    {
+        setTypeLeaf( cursor );
+        setKeyCount( cursor, 0 );
+        setRightSibling( cursor, NO_NODE_FLAG );
+        setLeftSibling( cursor, NO_NODE_FLAG );
+    }
 
-    void setTypeInternal( PageCursor cursor );
+    void initializeInternal( PageCursor cursor )
+    {
+        setTypeInternal( cursor );
+        setKeyCount( cursor, 0 );
+        setRightSibling( cursor, NO_NODE_FLAG );
+        setLeftSibling( cursor, NO_NODE_FLAG );
+    }
 
-    void setKeyCount( PageCursor cursor, int count );
+    // HEADER METHODS
 
-    void setRightSibling( PageCursor cursor, long rightSiblingId );
+    boolean isLeaf( PageCursor cursor )
+    {
+        return cursor.getByte( BYTE_POS_TYPE ) == LEAF_FLAG;
+    }
 
-    void setLeftSibling( PageCursor cursor, long leftSiblingId );
+    boolean isInternal( PageCursor cursor )
+    {
+        return cursor.getByte( BYTE_POS_TYPE ) == INTERNAL_FLAG;
+    }
 
-    KEY keyAt( PageCursor cursor, KEY into, int pos );
+    int keyCount( PageCursor cursor )
+    {
+        return cursor.getInt( BYTE_POS_KEYCOUNT );
+    }
 
-    void insertKeyAt( PageCursor cursor, KEY key, int pos, int keyCount, byte[] tmp );
+    long rightSibling( PageCursor cursor )
+    {
+        return cursor.getLong( BYTE_POS_RIGHTSIBLING );
+    }
 
-    void removeKeyAt( PageCursor cursor, int pos, byte[] tmp );
+    long leftSibling( PageCursor cursor )
+    {
+        return cursor.getLong( BYTE_POS_LEFTSIBLING );
+    }
 
-    // no setKeyAt since we only insert or remove keys, when they're there they're not changed
+    private void setTypeLeaf( PageCursor cursor )
+    {
+        cursor.putByte( BYTE_POS_TYPE, LEAF_FLAG );
+    }
 
-    VALUE valueAt( PageCursor cursor, VALUE value, int pos );
+    private void setTypeInternal( PageCursor cursor )
+    {
+        cursor.putByte( BYTE_POS_TYPE, INTERNAL_FLAG );
+    }
 
-    void insertValueAt( PageCursor cursor, VALUE value, int pos, int keyCount, byte[] tmp );
+    void setKeyCount( PageCursor cursor, int count )
+    {
+        cursor.putInt( BYTE_POS_KEYCOUNT, count );
+    }
 
-    void removeValueAt( PageCursor cursor, int pos, byte[] tmp );
+    void setRightSibling( PageCursor cursor, long rightSiblingId )
+    {
+        cursor.putLong( BYTE_POS_RIGHTSIBLING, rightSiblingId );
+    }
 
-    void setValueAt( PageCursor cursor, VALUE value, int pos );
+    void setLeftSibling( PageCursor cursor, long leftSiblingId )
+    {
+        cursor.putLong( BYTE_POS_LEFTSIBLING, leftSiblingId );
+    }
 
-    long childAt( PageCursor cursor, int pos );
+    // BODY METHODS
 
-    void insertChildAt( PageCursor cursor, long child, int pos, int keyCount, byte[] tmp );
+    KEY keyAt( PageCursor cursor, KEY into, int pos )
+    {
+        cursor.setOffset( keyOffset( pos ) );
+        layout.readKey( cursor, into );
+        return into;
+    }
 
-    void setChildAt( PageCursor cursor, long child, int pos );
+    void insertKeyAt( PageCursor cursor, KEY key, int pos, int keyCount, byte[] tmp )
+    {
+        insertSlotAt( cursor, pos, keyCount, keyOffset( 0 ), keySize, tmp );
+        cursor.setOffset( keyOffset( pos ) );
+        layout.writeKey( cursor, key );
+    }
 
-    // no removeChildAt since we don't do merge and stuff yet
+    void removeKeyAt( PageCursor cursor, int pos, byte[] tmp )
+    {
+        removeSlotAt( cursor, pos, keyOffset( 0 ), keySize, tmp );
+    }
 
-    int internalMaxKeyCount();
+    private void removeSlotAt( PageCursor cursor, int pos, int baseOffset, int itemSize, byte[] tmp )
+    {
+        int from = pos + 1;
+        int count = keyCount( cursor ) - from;
+        copyItems( cursor, from, count, baseOffset, itemSize, tmp );
+        writeItems( cursor, pos, count, baseOffset, itemSize, tmp );
+    }
 
-    int leafMaxKeyCount();
+    /**
+     * Moves items (key/value/child) one step to the right, which means rewriting all items of the particular type
+     * from pos - keyCount.
+     */
+    private void insertSlotAt( PageCursor cursor, int pos, int toExcluding, int baseOffset, int itemSize, byte[] tmp )
+    {
+        // Move all items after pos one step to the right
+        int count = toExcluding - pos;
+        if ( count > 0 )
+        {
+            copyItems( cursor, pos, count, baseOffset, itemSize, tmp );
+            writeItems( cursor, pos + 1, count, baseOffset, itemSize, tmp );
+        }
+    }
 
-    int keyOffset( int pos );
+    private void writeItems( PageCursor cursor, int pos, int count, int baseOffset, int itemSize, byte[] tmp )
+    {
+        cursor.setOffset( baseOffset + pos * itemSize );
+        cursor.putBytes( tmp, 0, count * itemSize );
+    }
 
-    int valueOffset( int pos );
+    private void copyItems( PageCursor cursor, int pos, int count, int baseOffset, int itemSize, byte[] tmp )
+    {
+        cursor.setOffset( baseOffset + pos * itemSize );
+        cursor.getBytes( tmp, 0, count * itemSize );
+    }
 
-    int childOffset( int pos );
+    VALUE valueAt( PageCursor cursor, VALUE value, int pos )
+    {
+        cursor.setOffset( valueOffset( pos ) );
+        layout.readValue( cursor, value );
+        return value;
+    }
 
-    boolean isNode( long node );
+    void insertValueAt( PageCursor cursor, VALUE value, int pos, int keyCount, byte[] tmp )
+    {
+        insertSlotAt( cursor, pos, keyCount, valueOffset( 0 ), valueSize, tmp );
+        setValueAt( cursor, value, pos );
+    }
 
-    int keySize();
+    void removeValueAt( PageCursor cursor, int pos, byte[] tmp )
+    {
+        removeSlotAt( cursor, pos, valueOffset( 0 ), valueSize, tmp );
+    }
 
-    int valueSize();
+    void setValueAt( PageCursor cursor, VALUE value, int pos )
+    {
+        cursor.setOffset( valueOffset( pos ) );
+        layout.writeValue( cursor, value );
+    }
 
-    int childSize();
+    long childAt( PageCursor cursor, int pos )
+    {
+        return cursor.getLong( childOffset( pos ) );
+    }
 
-    Comparator<KEY> keyComparator();
+    void insertChildAt( PageCursor cursor, long child, int pos, int keyCount, byte[] tmp )
+    {
+        insertSlotAt( cursor, pos, keyCount + 1, childOffset( 0 ), SIZE_CHILD, tmp );
+        setChildAt( cursor, child, pos );
+    }
 
-    // TODO: incremental methods towards V2
+    void setChildAt( PageCursor cursor, long child, int pos )
+    {
+        cursor.putLong( childOffset( pos ), child );
+    }
+
+    void writeChild( PageCursor cursor, long child )
+    {
+        cursor.putLong( child );
+    }
+
+    int internalMaxKeyCount()
+    {
+        return internalMaxKeyCount;
+    }
+
+    int leafMaxKeyCount()
+    {
+        return leafMaxKeyCount;
+    }
+
+    // HELPERS
+
+    int keyOffset( int pos )
+    {
+        return HEADER_LENGTH + pos * keySize;
+    }
+
+    private int valueOffset( int pos )
+    {
+        return HEADER_LENGTH + leafMaxKeyCount * keySize + pos * valueSize;
+    }
+
+    int childOffset( int pos )
+    {
+        return HEADER_LENGTH + internalMaxKeyCount * keySize + pos * SIZE_CHILD;
+    }
+
+    boolean isNode( long node )
+    {
+        return node != NO_NODE_FLAG;
+    }
+
+    int keySize()
+    {
+        return keySize;
+    }
+
+    int valueSize()
+    {
+        return valueSize;
+    }
+
+    int childSize()
+    {
+        return SIZE_CHILD;
+    }
+
+    Comparator<KEY> keyComparator()
+    {
+        return layout;
+    }
 
     int readKeysWithInsertRecordInPosition( PageCursor cursor, Consumer<PageCursor> newRecordWriter,
-            int insertPosition, int totalNumberOfRecords, byte[] into );
+            int insertPosition, int totalNumberOfRecords, byte[] into )
+    {
+        return readRecordsWithInsertRecordInPosition( cursor, newRecordWriter, insertPosition, totalNumberOfRecords,
+                keySize, keyOffset( 0 ), into );
+    }
 
     int readValuesWithInsertRecordInPosition( PageCursor cursor, Consumer<PageCursor> newRecordWriter,
-            int insertPosition, int totalNumberOfRecords, byte[] into );
+            int insertPosition, int totalNumberOfRecords, byte[] into )
+    {
+        return readRecordsWithInsertRecordInPosition( cursor, newRecordWriter, insertPosition, totalNumberOfRecords,
+                valueSize, valueOffset( 0 ), into );
+    }
 
     int readChildrenWithInsertRecordInPosition( PageCursor cursor, Consumer<PageCursor> newRecordWriter,
-            int insertPosition, int totalNumberOfRecords, byte[] into );
+            int insertPosition, int totalNumberOfRecords, byte[] into )
+    {
+        return readRecordsWithInsertRecordInPosition( cursor, newRecordWriter, insertPosition, totalNumberOfRecords,
+                childSize(), childOffset( 0 ), into );
+    }
 
-    void writeKeys( PageCursor cursor, byte[] source, int sourcePos, int targetPos, int count );
+    /**
+     * Leaves cursor on same page as when called. No guarantees on offset.
+     *
+     * Create a byte[] with totalNumberOfRecords of recordSize from cursor reading from baseRecordOffset
+     * with newRecord inserted in insertPosition, with the following records shifted to the right.
+     *
+     * Simply: Records of size recordSize that can be read from offset baseRecordOffset in page pinned by cursor has
+     * some ordering. This ordering is preserved with new record inserted in insertPosition in the returned byte[],
+     * NOT in the page.
+     *
+     * @param cursor                {@link org.neo4j.io.pagecache.PageCursor} pinned to page to read records from
+     * @param newRecordWriter       new data to be inserted in insertPosition in returned byte[].
+     *                              This is a {@link Consumer} since the type of data can differ (value/child),
+     *                              although perhaps this can be done in a better way
+     * @param insertPosition        position of newRecord. 0 being before all other records,
+     *                              (totalNumberOfRecords - 1) being after all other records
+     * @param totalNumberOfRecords  the total number of records to be contained in returned byte[], including newRecord
+     * @param recordSize            the size in number of bytes of one record
+     * @param baseRecordOffset      the offset from where cursor should start read records
+     * @param into                  byte array to copy bytes into
+     * @return                      number of bytes copied into the {@code into} byte[],
+     *                              that is insertPosition * recordSize
+     */
+    private int readRecordsWithInsertRecordInPosition( PageCursor cursor, Consumer<PageCursor> newRecordWriter,
+            int insertPosition, int totalNumberOfRecords, int recordSize, int baseRecordOffset, byte[] into )
+    {
+        int length = (totalNumberOfRecords) * recordSize;
 
-    void writeValues( PageCursor cursor, byte[] source, int sourcePos, int targetPos, int count );
+        // First read all records
 
-    void writeChildren( PageCursor cursor, byte[] source, int sourcePos, int targetPos, int count );
+        // Read all records on previous to insertPosition
+        cursor.setOffset( baseRecordOffset );
+        cursor.getBytes( into, 0, insertPosition * recordSize );
 
-    void writeChild( PageCursor cursor, long child );
+        // Read newRecord
+        // TODO: A bit expensive to wrap in a PageCursor tin foil just to write this middle key
+        PageCursor buffer = ByteArrayPageCursor.wrap( into, insertPosition * recordSize, recordSize );
+        newRecordWriter.accept( buffer );
+
+        // Read all records following insertPosition
+        cursor.setOffset( baseRecordOffset + insertPosition * recordSize );
+        cursor.getBytes( into, (insertPosition + 1) * recordSize,
+                ((totalNumberOfRecords - 1) - insertPosition) * recordSize );
+        return length;
+    }
+
+    private void writeAll( PageCursor cursor, byte[] source, int sourcePos, int targetPos, int count,
+            int baseOffset, int recordSize )
+    {
+        int arrayOffset = sourcePos * recordSize;
+        cursor.setOffset( baseOffset + recordSize * targetPos );
+        cursor.putBytes( source, arrayOffset, count * recordSize );
+    }
+
+    void writeKeys( PageCursor cursor, byte[] source, int sourcePos, int targetPos, int count )
+    {
+        writeAll( cursor, source, sourcePos, targetPos, count, keyOffset( 0 ), keySize() );
+    }
+
+    void writeValues( PageCursor cursor, byte[] source, int sourcePos, int targetPos, int count )
+    {
+        writeAll( cursor, source, sourcePos, targetPos, count, valueOffset( 0 ), valueSize() );
+    }
+
+    void writeChildren( PageCursor cursor, byte[] source, int sourcePos, int targetPos, int count )
+    {
+        writeAll( cursor, source, sourcePos, targetPos, count, childOffset( 0 ), childSize() );
+    }
 }
